Move templated function definitions in header.
