+++ queso/embedding/trimmed_domain.cpp
@@ -77,21 +77,6 @@ const BoundingBoxType TrimmedDomain::GetBoundingBoxOfTrimmedDomain() const {
     return bounding_box;
 }
 
-template<typename BoundaryIntegrationPointType>
-Unique<std::vector<BoundaryIntegrationPointType>> TrimmedDomain::pGetBoundaryIps() const{
-    // Pointer to boundary integration points
-    auto p_boundary_ips = MakeUnique<std::vector<BoundaryIntegrationPointType>>();
-
-    p_boundary_ips->reserve(mpTriangleMesh->NumOfTriangles()*12UL);
-    for( IndexType triangle_id = 0; triangle_id < mpTriangleMesh->NumOfTriangles(); ++triangle_id ){
-        IndexType method = 3; // Creates 12 points per triangle.
-        auto p_new_points = mpTriangleMesh->pGetIPsGlobal<BoundaryIntegrationPointType>(triangle_id, method);
-        p_boundary_ips->insert(p_boundary_ips->end(), p_new_points->begin(), p_new_points->end());
-    }
-
-    return p_boundary_ips;
-}
-
 IntersectionStatusType TrimmedDomain::GetIntersectionState(
         const PointType& rLowerBound, const PointType& rUpperBound, double Tolerance) const
 {
@@ -106,8 +91,5 @@ IntersectionStatusType TrimmedDomain::GetIntersectionState(
     // If triangle is not intersected, center location will determine if inside or outside.
     return status;
 }
-
-/// Explicit function instantiation
-template Unique<std::vector<BoundaryIntegrationPoint>> TrimmedDomain::pGetBoundaryIps<BoundaryIntegrationPoint>() const;
 
 } // End namespace queso
