+++ queso/io/io_utilities.cpp
@@ -206,325 +206,6 @@ bool IO::WriteDisplacementToVTK(const std::vector<Vector3d>& rDisplacement,
     return true;
 }
 
-template<typename TElementType>
-bool IO::WriteElementsToVTK(const ElementContainer<TElementType>& rElementContainer, //PolygonMesh
-                            const char* Filename,
-                            const bool Binary){
-
-    const SizeType num_elements = rElementContainer.size();
-
-    std::ofstream file;
-    if(Binary)
-        file.open(Filename, std::ios::out | std::ios::binary);
-    else
-        file.open(Filename);
-
-    file << "# vtk DataFile Version 4.1" << std::endl;
-    file << "vtk output" << std::endl;
-    if(Binary)
-        file << "BINARY"<< std::endl;
-    else
-        file << "ASCII"<< std::endl;
-
-    file << "DATASET UNSTRUCTURED_GRID" << std::endl;
-    file << "POINTS " << num_elements*8 << " double" << std::endl;
-
-    const auto begin_el_itr = rElementContainer.begin();
-    for( IndexType i = 0; i < rElementContainer.size(); ++i){
-        const auto& el_itr = *(begin_el_itr + i);
-        const auto& lower_point = el_itr->GetBoundsXYZ().first;
-        const auto& upper_point = el_itr->GetBoundsXYZ().second;
-
-        if( Binary ){
-            double rx0 = lower_point[0];
-            SwapEnd(rx0);
-            double rx1 = upper_point[0];
-            SwapEnd(rx1);
-            double ry0 = lower_point[1];
-            SwapEnd(ry0);
-            double ry1 = upper_point[1];
-            SwapEnd(ry1);
-            double rz0 = lower_point[2];
-            SwapEnd(rz0);
-            double rz1 = upper_point[2];
-            SwapEnd(rz1);
-
-            file.write(reinterpret_cast<char*>(&rx0), sizeof(double));
-            file.write(reinterpret_cast<char*>(&ry0), sizeof(double));
-            file.write(reinterpret_cast<char*>(&rz0), sizeof(double));
-
-            file.write(reinterpret_cast<char*>(&rx1), sizeof(double));
-            file.write(reinterpret_cast<char*>(&ry0), sizeof(double));
-            file.write(reinterpret_cast<char*>(&rz0), sizeof(double));
-
-            file.write(reinterpret_cast<char*>(&rx1), sizeof(double));
-            file.write(reinterpret_cast<char*>(&ry1), sizeof(double));
-            file.write(reinterpret_cast<char*>(&rz0), sizeof(double));
-
-            file.write(reinterpret_cast<char*>(&rx0), sizeof(double));
-            file.write(reinterpret_cast<char*>(&ry1), sizeof(double));
-            file.write(reinterpret_cast<char*>(&rz0), sizeof(double));
-
-            file.write(reinterpret_cast<char*>(&rx0), sizeof(double));
-            file.write(reinterpret_cast<char*>(&ry0), sizeof(double));
-            file.write(reinterpret_cast<char*>(&rz1), sizeof(double));
-
-            file.write(reinterpret_cast<char*>(&rx1), sizeof(double));
-            file.write(reinterpret_cast<char*>(&ry0), sizeof(double));
-            file.write(reinterpret_cast<char*>(&rz1), sizeof(double));
-
-            file.write(reinterpret_cast<char*>(&rx1), sizeof(double));
-            file.write(reinterpret_cast<char*>(&ry1), sizeof(double));
-            file.write(reinterpret_cast<char*>(&rz1), sizeof(double));
-
-            file.write(reinterpret_cast<char*>(&rx0), sizeof(double));
-            file.write(reinterpret_cast<char*>(&ry1), sizeof(double));
-            file.write(reinterpret_cast<char*>(&rz1), sizeof(double));
-        }
-        else {
-            file << lower_point[0] << ' ' << lower_point[1] << ' ' << lower_point[2] << std::endl;
-            file << upper_point[0] << ' ' << lower_point[1] << ' ' << lower_point[2] << std::endl;
-            file << upper_point[0] << ' ' << upper_point[1] << ' ' << lower_point[2] << std::endl;
-            file << lower_point[0] << ' ' << upper_point[1] << ' ' << lower_point[2] << std::endl;
-            file << lower_point[0] << ' ' << lower_point[1] << ' ' << upper_point[2] << std::endl;
-            file << upper_point[0] << ' ' << lower_point[1] << ' ' << upper_point[2] << std::endl;
-            file << upper_point[0] << ' ' << upper_point[1] << ' ' << upper_point[2] << std::endl;
-            file << lower_point[0] << ' ' << upper_point[1] << ' ' << upper_point[2] << std::endl;
-        }
-    }
-    file << std::endl;
-    // Write Cells
-    file << "Cells " << num_elements << " " << num_elements*9 << std::endl;
-    for( int i = 0; i < static_cast<int>(rElementContainer.size()); ++i){
-        if( Binary ){
-            int k = 8;
-            WriteBinary(file, k);
-        for( int j = 0; j < 8; ++j){
-            k = 8*i+j;
-            WriteBinary(file, k);
-        }
-        }
-        else {
-            file << 8 << ' ' << 8*i     << ' ' << 8*i + 1 << ' ' << 8*i + 2 << ' ' << 8*i + 3
-                        << ' ' << 8*i + 4 << ' ' << 8*i + 5 << ' ' << 8*i + 6 << ' ' << 8*i + 7 << std::endl;
-        }
-    }
-    file << std::endl;
-
-    file << "CELL_TYPES " << rElementContainer.size() << std::endl;
-    for( int i = 0; i < static_cast<int>(rElementContainer.size()); ++i){
-        if( Binary ){
-            int k = 12;
-            WriteBinary(file, k);
-        }
-        else {
-            file << 12 << std::endl;
-        }
-    }
-    file << std::endl;
-    file.close();
-
-    return true;
-}
-
-template<typename TElementType>
-bool IO::WritePointsToVTK(const ElementContainer<TElementType>& rElementContainer,
-                          const char* type,
-                          const char* Filename,
-                          const bool Binary){
-
-    QuESo_ERROR_IF( std::string(type) != "All") << "Only integration points option 'All' is available. \n";
-
-    const auto& p_points = rElementContainer.pGetPoints(type);
-    const IndexType num_points = p_points->size();
-    const IndexType num_elements = p_points->size();
-
-    std::ofstream file;
-    if(Binary)
-        file.open(Filename, std::ios::out | std::ios::binary);
-    else
-        file.open(Filename);
-
-    file << "# vtk DataFile Version 4.1" << std::endl;
-    file << "vtk output" << std::endl;
-    if(Binary)
-        file << "BINARY"<< std::endl;
-    else
-        file << "ASCII"<< std::endl;
-
-
-    file << "DATASET UNSTRUCTURED_GRID" << std::endl;
-    file << "POINTS " << num_points << " double" << std::endl;
-
-    const auto el_it_ptr_begin = rElementContainer.begin();
-    for( IndexType i = 0; i < rElementContainer.size(); ++i){
-        const auto& el_ptr = (*(el_it_ptr_begin + i));
-        const auto& r_points = el_ptr->GetIntegrationPoints();
-        for( const auto& r_point : r_points ){
-            auto point_global = el_ptr->PointFromParamToGlobal(r_point.data());
-
-            if( Binary ){
-                WriteBinary(file, point_global[0]);
-                WriteBinary(file, point_global[1]);
-                WriteBinary(file, point_global[2]);
-            }
-            else {
-                file << point_global[0] << ' ' << point_global[1] << ' ' << point_global[2] << std::endl;
-            }
-        }
-
-    }
-    file << std::endl;
-
-    //Write Cells
-    file << "Cells " << num_elements << " " << num_elements*2 << std::endl;
-    for( IndexType i = 0; i < num_elements; ++i){
-        if( Binary ){
-            int k = 1;
-            WriteBinary(file, k);
-            k = i;
-            WriteBinary(file, k);
-        }
-        else {
-            file << 1 << ' ' << i << std::endl;
-        }
-    }
-    file << std::endl;
-
-    file << "CELL_TYPES " << num_elements << std::endl;
-    for( IndexType i = 0; i < num_elements; ++i){
-        if( Binary ){
-            int k = 1;
-            WriteBinary(file, k);
-        }
-        else {
-            file << 1 << std::endl;
-        }
-    }
-    file << std::endl;
-
-    file << "POINT_DATA " << num_points << std::endl;
-    file << "SCALARS Weights double 1" << std::endl;
-    file << "LOOKUP_TABLE default" << std::endl;
-    for( IndexType i = 0; i < rElementContainer.size(); ++i){
-        const auto& el_ptr = (*(el_it_ptr_begin + i));
-        const auto& points = el_ptr->GetIntegrationPoints();
-        for( const auto& point : points ){
-            if( Binary ){
-                double rw = point.Weight();
-                WriteBinary(file, rw);
-            }
-            else {
-                file << point.Weight() << std::endl;
-            }
-        }
-    }
-    file << std::endl;
-    file.close();
-
-    return true;
-}
-
-
-template<typename Type>
-bool IO::WritePointsToVTK(const std::vector<Type>& rPoints,
-                          const char* Filename,
-                          const bool Binary){
-
-    const auto begin_points_it_ptr = rPoints.begin();
-    const IndexType num_points = rPoints.size();
-    const IndexType num_elements = rPoints.size();
-
-    std::ofstream file;
-    if(Binary)
-        file.open(Filename, std::ios::out | std::ios::binary);
-    else
-        file.open(Filename);
-
-    file << "# vtk DataFile Version 4.1" << std::endl;
-    file << "vtk output" << std::endl;
-    if(Binary)
-        file << "BINARY"<< std::endl;
-    else
-        file << "ASCII"<< std::endl;
-
-
-    file << "DATASET UNSTRUCTURED_GRID" << std::endl;
-    file << "POINTS " << num_points << " double" << std::endl;
-
-    for(IndexType i = 0; i < num_points; ++i){
-        auto points_it = (begin_points_it_ptr + i);
-        if( Binary ){
-            // Make sure to create copy of points. "WriteBinary" will change them.
-            auto p_x = (*points_it)[0];
-            auto p_y = (*points_it)[1];
-            auto p_z = (*points_it)[2];
-            WriteBinary(file, p_x);
-            WriteBinary(file, p_y);
-            WriteBinary(file, p_z);
-        }
-        else {
-            file << (*points_it)[0] << ' ' << (*points_it)[1] << ' ' << (*points_it)[2] << std::endl;
-        }
-    }
-    file << std::endl;
-
-    //Write Cells
-    file << "Cells " << num_elements << " " << num_elements*2 << std::endl;
-    for( IndexType i = 0; i < num_elements; ++i){
-        if( Binary ){
-            int k = 1;
-            WriteBinary(file, k);
-            k = i;
-            WriteBinary(file, k);
-        }
-        else {
-            file << 1 << ' ' << i << std::endl;
-        }
-    }
-    file << std::endl;
-
-    file << "CELL_TYPES " << num_elements << std::endl;
-    for( IndexType i = 0; i < num_elements; ++i){
-        if( Binary ){
-            int k = 1;
-            WriteBinary(file, k);
-        }
-        else {
-        file << 1 << std::endl;
-        }
-    }
-    file << std::endl;
-
-    file << "POINT_DATA " << num_points << std::endl;
-    file << "SCALARS Weights double 1" << std::endl;
-    file << "LOOKUP_TABLE default" << std::endl;
-    for(IndexType i = 0; i < num_points; ++i){
-        auto points_it = (begin_points_it_ptr + i);
-
-        if( Binary ){
-            double rw = points_it->Weight();
-            WriteBinary(file, rw);
-        }
-        else {
-            file << points_it->Weight() << std::endl;
-        }
-    }
-    file << std::endl;
-    file.close();
-
-    return true;
-}
-
-// Instantiation
-template bool IO::WritePointsToVTK<IntegrationPoint>(const std::vector<IntegrationPoint>& rPoints,
-                                                    const char* Filename,
-                                                    const bool Binary);
-
-template bool IO::WritePointsToVTK<BoundaryIntegrationPoint>(const std::vector<BoundaryIntegrationPoint>& rPoints,
-                                                            const char* Filename,
-                                                            const bool Binary);
-
 ////// Private member functions //////
 
 bool IO::STLIsInASCIIFormat(const char* Filename) {
@@ -734,17 +415,6 @@ bool IO::ReadMeshFromSTL_Binary(TriangleMeshInterface& rTriangleMesh,
     return rTriangleMesh.Check();
 }
 
-//// Explicit function instantiation
-template bool IO::WritePointsToVTK<Element<IntegrationPoint, BoundaryIntegrationPoint>>(
-                const ElementContainer<Element<IntegrationPoint, BoundaryIntegrationPoint>>& rElementContainer,
-                const char* type,
-                const char* Filename,
-                const bool Binary);
-
-template bool IO::WriteElementsToVTK<Element<IntegrationPoint, BoundaryIntegrationPoint>>(
-                const ElementContainer<Element<IntegrationPoint, BoundaryIntegrationPoint>>& rElementContainer,
-                const char* Filename,
-                const bool Binary);
 
 } // End namespace queso
